Send nextFrame to renderer
About the HACK next to the send to nextFrame:
I have warnings when I try to access a IO register the emulator does not
know about (ie. sound registers), I started the emulation and at the
same time loaded a video showing a GameBoy booting, the warnings were
synchronized with the classic boot sound.

I'm happy.