Fix LD A,16b
While I did have a bug, trying to fix it led me astray, for my emulation
was actually good enough to do most of what the ROM told me to, and the
ROM told me to jump into data and then straight into a 0xFF (RST 38) loop.

The ROM I was using was bad.

I replaced the ROM, went into another RST38 loop and saw that my LD was
flawed.